requirements.txt
Django==2.0.5
django-filter==1.1.0
djangorestframework==3.8.2
Markdown==2.6.11
pytz==2018.4
gunicorn==19.8.1
psycopg2==2.7.1
dj-database-url==0.5.0
whitenoise==3.3.1
requests==2.18.4
